+++ package.json
@@ -57,5 +57,8 @@
     "ts-node": "10.4.0",
     "typescript": "4.0.5"
   },
-  "packageManager": "yarn@3.2.1"
+  "packageManager": "yarn@3.2.1",
+  "resolutions": {
+    "@graasp/sdk": "portal:/Users/kim/git/graasp-utils"
+  }
 }
