+++ src/index.ts
@@ -1,4 +1,6 @@
 export { default } from './plugin';
+export { default as itemPlugin } from './itemPlugin';
+// export { default } from './memplugin';
 export { ActionTaskManager } from './services/action/task-manager';
 export { ActionService } from './services/action/db-service';
 export { BaseAction } from './services/action/base-action';
